Run evals

+++ target_benchmark/retrievers/llama_index/embedding_utils.py
@@ -1,12 +1,12 @@
 import json
 import re
 from pathlib import Path
-from typing import Union
+from typing import Dict, Union
 
 import pandas as pd
 from openai import OpenAI
 from pydantic import BaseModel, Field
-from sqlalchemy import Column, Engine, Integer, MetaData, String, Table
+from sqlalchemy import Column, Engine, MetaData, String, Table
 
 
 class TableInfo(BaseModel):
@@ -25,6 +25,10 @@ Give me a summary of the table with the following JSON format.
 
 - The table name must be unique to the table and describe it while being concise.
 - Do NOT output a generic table name (e.g. table, my_table).
+- You are allowed to output a name like <meaningful_table_name>_2, since some big tables are broken up into many smaller tables.
+
+Existing Table Names:
+{names_tried}
 
 Table:
 {table_str}
@@ -34,10 +38,10 @@ Summary: """
 client = OpenAI()
 
 
-def _get_table_info_with_index(
-    table_info_dir: str, table_name: str
+def get_table_info_with_index(
+    table_info_dir: str, db_id: str, table_name: str
 ) -> Union[TableInfo, None]:
-    results_gen = Path(table_info_dir).glob(f"{table_name}_*")
+    results_gen = Path(table_info_dir).glob(f"{db_id}_{table_name}.json")
     results_list = list(results_gen)
     if len(results_list) == 0:
         return None
@@ -49,34 +53,61 @@ def _get_table_info_with_index(
 
 
 def construct_table_info(
-    table_info_dir: str, df: pd.DataFrame, database_id: str, table_name: str
-) -> TableInfo:
-    table_info = _get_table_info_with_index(table_info_dir, table_name)
-    if table_info:
+    table_info_dir: str,
+    df: pd.DataFrame,
+    database_id: str,
+    table_name: str,
+    existing_names: Dict,
+) -> Union[TableInfo, None]:
+    cleaned_table_name = re.sub(r"[^a-zA-Z0-9_]", "_", table_name)
+    if isinstance(database_id, int):
+        database_id = str(database_id)
+    cleaned_db_id = re.sub(r"[^a-zA-Z0-9_]", "_", database_id)
+    table_info = get_table_info_with_index(
+        table_info_dir, database_id, cleaned_table_name
+    )
+
+    # check if
+    # - table info has been constructed already
+    # - and not appeared in other tables
+    if table_info and table_info.table_name not in existing_names:
         return table_info
 
     df_str = df.head(10).to_csv()
-    table_info_completion = client.beta.chat.completions.parse(
-        model="gpt-4o-2024-08-06",
-        messages=[
-            {"role": "system", "content": "You are a helpful AI assistant."},
-            {"role": "user", "content": prompt_str.format(table_str=df_str)},
-        ],
-        response_format=TableInfo,
-    )
-    message = table_info_completion.choices[0].message
-    if not message.parsed:
-        raise json.JSONDecodeError
-    table_info = TableInfo(
-        table_name=message.parsed.table_name,
-        table_summary=message.parsed.table_summary,
-    )
+    names_tried = set()
+    for i in range(15):  # try up to 15 times
+        table_info_completion = client.beta.chat.completions.parse(
+            model="gpt-4o-2024-08-06",
+            messages=[
+                {"role": "system", "content": "You are a helpful AI assistant."},
+                {
+                    "role": "user",
+                    "content": prompt_str.format(
+                        table_str=df_str, names_tried=names_tried
+                    ),
+                },
+            ],
+            response_format=TableInfo,
+        )
+        message = table_info_completion.choices[0].message
+        if not message.parsed:
+            raise json.JSONDecodeError
+        if message.parsed.table_name in existing_names:
+            # duplicate table names, try again
+            names_tried.add(message.parsed.table_name)
+            continue
+        table_info = TableInfo(
+            table_name=message.parsed.table_name,
+            table_summary=message.parsed.table_summary,
+        )
+        out_file_path = (
+            Path(table_info_dir) / f"{cleaned_db_id}_{cleaned_table_name}.json"
+        )
+        with open(out_file_path, "w") as file:
+            json.dump(table_info.model_dump(), file)
 
-    table_info.table_name = f"{database_id}:{table_name}:{table_info.table_name}"  # forcefully prepend the official table name
-    out_file_path = f"{table_info_dir}/{database_id}_{table_name}.json"
-    with open(out_file_path, "w") as file:
-        json.dump(table_info.model_dump(), file)
-    return table_info
+        return table_info
+    return None
 
 
 # Function to create a sanitized column name
@@ -94,20 +125,26 @@ def create_table_from_dataframe(
     df = df.rename(columns=sanitized_columns)
 
     # Dynamically create columns based on DataFrame columns and data types
-    columns = [
-        Column(col, String if dtype == "object" else Integer)
-        for col, dtype in zip(df.columns, df.dtypes)
-    ]
+    # columns = [
+    #     Column(col, String if dtype == "object" else Integer)
+    #     for col, dtype in zip(df.columns, df.dtypes)
+    # ]
+    columns = [Column("Column", String)]
+    # Column(
+    #     df.columns.values[0], String if df.dtypes.values[0] == "object" else Integer
+    # )
+    # ]
 
     # Create a table with the defined columns
-    table = Table(table_name, metadata_obj, *columns)
+    Table(table_name, metadata_obj, *columns)
 
     # Create the table in the database
     metadata_obj.create_all(engine)
 
     # Insert data from DataFrame into the table
-    with engine.connect() as conn:
-        for _, row in df.iterrows():
-            insert_stmt = table.insert().values(**row.to_dict())
-            conn.execute(insert_stmt)
-        conn.commit()
+    # with engine.connect() as conn:
+    #     for _, row in df.iterrows():
+    #         insert_stmt = table.insert().values(row.to_dict()[df.columns[0]])
+    #         conn.execute(insert_stmt)
+    #         break
+    #     conn.commit()
